// src/Organizer.SPA/src/app/api/OrganizerApiClient/model/customer.ts
/**
 *  Organizer
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Customer { 
    id?: number;
    name?: string;
    shortDescription?: string;
    color?: string;
    textColor?: string;
    projectColors?: string;
    referral?: string;
    address?: string;
    note?: string;
}